refactor(diagnosis): 归位诊断契约与生命周期逻辑
