+++ src/diagnosis/contracts.ts
