+++ src/types/index.ts
@@ -1,4 +1,3 @@
 export * from './shared.js';
 export * from './doctor.js';
-export * from './diagnosis.js';
 export * from './skill-index.js';
