+++ test/scripts/legacy-evaluation-removal.test.ts
@@ -8,6 +8,7 @@ const FORBIDDEN_FILES = [
   'src/types/storage.ts',
   'src/types/eval.ts',
   'src/types/judge.ts',
+  'src/types/diagnosis.ts',
   'src/eval-workflows/run-evaluation.ts',
   'src/eval-workflows/evaluation-pipeline.ts',
   'src/eval-workflows/batch-evaluation-workflow.ts',
@@ -25,6 +26,7 @@ const FORBIDDEN_SPECIFIER_FRAGMENTS = [
   '/types/storage',
   '/types/eval',
   '/types/judge',
+  '/types/diagnosis',
   '/eval-workflows/run-evaluation',
   '/eval-workflows/evaluation-pipeline',
   '/server/report-store',
